Trying to get D9 build on Marlin 2.0.4.4. LEVEL_BED_CORNERS is an issue
Having a conflict with the old implementation of the creality_DWIN
screen used. Tinkering with this while looking hopefully not
breaking the extensible UI.

Need to fix this so that the Macro for this is correct for the inset
manual levelling features or remove it entirely.

Bit of a PITA.